requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Jinja2==3.1.2
paramiko==3.1.0
pycodestyle==2.10.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
scp==0.14.5